requirements/base.txt
Django==4.0
django-cors-headers==3.10.1
django-cuser==2017.3.16
django-extensions==3.1.5
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2-binary==2.9.3
